// src/ExecutionInput/enums.ts
export enum FieldTypes {
  LABEL = "LABEL",
  SMALL_TEXT = "SMALL_TEXT",
  LARGE_TEXT = "LARGE_TEXT",
  NUMBER = "NUMBER",
  PHONE = "PHONE",
  DATE = "DATE",
  TIME = "TIME",
  SSN = "SSN",
  SELECT = 'SELECT',
  DATE_RANGE = "DATE_RANGE",
  RADIO_BUTTON = 'RADIO_BUTTON',
  RADIO_BUTTON_DESCRIPTION = 'RADIO_BUTTON_DESCRIPTION',
  EMAIL = "EMAIL",
  LINK = "LINK",
  SCHEDULE = "SCHEDULE",
  BUTTON = "BUTTON",
  RADIO_BUTTON_IMAGE = "RADIO_BUTTON_IMAGE"
}